Relocate cast_record_batch into its own module to shed the datafusion dependency
A dependency from optimize on the cast_record_batch function was added
which cannot be met without the `datafusion` feature enabled

See #1926
